+++ test/test_m2m_iscope.py
@@ -10,6 +10,7 @@ import csv
 import os
 import shutil
 import subprocess
+import sys
 
 from metage2metabo import utils
 
@@ -640,7 +641,10 @@ def test_m2m_iscope_call():
         assert EXPECTED_PRODUCED_COMPOUNDS[compound] ==  named_compounds[compound]
 
     # clean
-    shutil.rmtree(respath)
+    # Due to unstable behaviour of os.unlink on Windows, do not delete the file.
+    # Refer to: https://github.com/python/cpython/issues/109608
+    if sys.platform != 'win32':
+        shutil.rmtree(respath)
 
 if __name__ == "__main__":
     test_m2m_iscope_call()
